Fix -t flag for setting timestamp
The flag never had any effect because the error-checking condition was
inverted. It set the timestamp only when `err != nil`, so the timestamp
was always a zero value.

This also changes honeyvent to exit with a non-zero status code when
time-parsing fails rather than silently continuing with the current
time.